+++ tests/testthat/test-tbl_summary.R
@@ -1,3 +1,4 @@
+# tbl_summary(data) ------------------------------------------------------------
 test_that("tbl_summary(data)", {
   # creates table when data frame is passed
   expect_snapshot(tbl_summary(data = trial) |> as.data.frame())
@@ -12,7 +13,7 @@ test_that("tbl_summary(data) errors properly", {
   expect_snapshot(error = TRUE, tbl_summary(data = dplyr::tibble()))
 })
 
-
+# tbl_summary(by) --------------------------------------------------------------
 test_that("tbl_summary(by)", {
   expect_snapshot(tbl_summary(data = trial, by = trt) |> as.data.frame())
   expect_snapshot(tbl_summary(data = mtcars, by = am) |> as.data.frame())
@@ -24,6 +25,7 @@ test_that("tbl_summary(by) errors properly", {
   expect_snapshot(error = TRUE, tbl_summary(mtcars, by = c("mpg", "am")))
 })
 
+# tbl_summary(label) -----------------------------------------------------------
 test_that("tbl_summary(label)", {
   expect_error(
     tbl <- tbl_summary(
@@ -56,6 +58,7 @@ test_that("tbl_summary(label) errors properly", {
   )
 })
 
+# tbl_summary(statistic) -------------------------------------------------------
 test_that("tbl_summary(statistic)", {
   # categorical summary
   expect_equal(
@@ -122,7 +125,221 @@ test_that("tbl_summary(statistic) errors properly", {
   )
 })
 
+# tbl_summary(digit) -----------------------------------------------------------
+test_that("tbl_summary(digit)", {
+  expect_error(
+    tbl <- tbl_summary(
+      trial,
+      include = c(age, response, marker, ttdeath),
+      digits = list(
+        # using named list to change 2 of the 3 statistics
+        age = list(median = 4, p25 = \(x) style_number(x, digits = 2)),
+        # using a vector of integers
+        response = c(0, 3),
+        # using a single integer that will apply to all stats
+        marker = 0,
+        # passing a single function that will apply to all stats
+        ttdeath = list(\(x) style_number(x, digits = 2))
+      ),
+      missing = "no"
+    ) |>
+      modify_column_unhide(variable) |>
+      as.data.frame(col_labels = FALSE),
+    NA
+  )
 
+  # check the correct stats
+  expect_equal(
+    tbl |>
+      dplyr::filter(variable == "age") |>
+      dplyr::pull(stat_0),
+    "47.0000 (38.00, 57)"
+  )
+
+  expect_equal(
+    tbl |>
+      dplyr::filter(variable == "response") |>
+      dplyr::pull(stat_0),
+    "61 (31.606%)"
+  )
+
+  expect_equal(
+    tbl |>
+      dplyr::filter(variable == "marker") |>
+      dplyr::pull(stat_0),
+    "1 (0, 1)"
+  )
+
+  expect_equal(
+    tbl |>
+      dplyr::filter(variable == "ttdeath") |>
+      dplyr::pull(stat_0),
+    "22.41 (15.92, 24.00)"
+  )
+})
+
+test_that("tbl_summary(digit) errors properly", {
+  expect_error(
+    tbl_summary(
+      trial,
+      include = age,
+      digits = list(
+        age = list(median = letters, # this is not a function!
+                   p25 = \(x) style_number(x, digits = 2))
+      ),
+      missing = "no"
+    ),
+    "*"
+  )
+})
+
+# tbl_summary(type) ------------------------------------------------------------
+test_that("tbl_summary(type)", {
+  expect_snapshot(
+    tbl_summary(
+      trial,
+      include = c(age, marker, response, stage),
+      type = list(age = "continuous", marker = "continuous2", response = "dichotomous", state = "categorical"),
+      missing = "no"
+    ) |>
+      getElement("table_body") |>
+      dplyr::select(variable, summary_type, row_type, label)
+  )
+
+  # can use the default type to select variables to change the summary type
+  expect_equal(
+    tbl_summary(
+      trial,
+      type = list(all_continuous() ~ "continuous2", all_dichotomous() ~ "continuous"),
+      include = c(age, marker, response),
+      missing = "no"
+    ) |>
+      getElement("inputs") |>
+      getElement("type"),
+    list(age = "continuous2", marker = "continuous2", response = "continuous")
+  )
+
+  # yes/no variables default to dichotomous
+  expect_equal(
+    data.frame(yn = c("no", "yes", "yes")) |>
+      tbl_summary() |>
+      getElement("inputs") |>
+      getElement("value") |>
+      getElement("yn"),
+    "yes"
+  )
+  expect_equal(
+    data.frame(
+      yn = c("no", "yes", "yes") |> factor()
+    ) |>
+      tbl_summary() |>
+      getElement("inputs") |>
+      getElement("value") |>
+      getElement("yn"),
+    "yes"
+  )
+  expect_equal(
+    data.frame(
+      yn = c("no", "yes", "yes") |> factor(levels = c("yes", "no"))
+    ) |>
+      tbl_summary() |>
+      getElement("inputs") |>
+      getElement("value") |>
+      getElement("yn"),
+    "yes"
+  )
+  expect_equal(
+    data.frame(
+      yn = c("no", "no", "no") |> factor(levels = c("no", "yes"))
+    ) |>
+      tbl_summary() |>
+      getElement("inputs") |>
+      getElement("value") |>
+      getElement("yn"),
+    "yes"
+  )
+
+  # a yes or no only character defaults to categorical
+  expect_equal(
+    data.frame(yn = c("yes", "yes")) |>
+      tbl_summary() |>
+      getElement("inputs") |>
+      getElement("type") |>
+      getElement("yn"),
+    "categorical"
+  )
+  expect_equal(
+    data.frame(yn = c("no", "no")) |>
+      tbl_summary() |>
+      getElement("inputs") |>
+      getElement("type") |>
+      getElement("yn"),
+    "categorical"
+  )
+  expect_equal(
+    data.frame(yn = c("nO", "yEs", "yEs")) |>
+      tbl_summary() |>
+      getElement("inputs") |>
+      getElement("value") |>
+      getElement("yn"),
+    "yEs"
+  )
+
+  # a zero or one only numeric defaults to categorical
+  expect_equal(
+    data.frame(yn = c(0, 0)) |>
+      tbl_summary() |>
+      getElement("inputs") |>
+      getElement("type") |>
+      getElement("yn"),
+    "categorical"
+  )
+  expect_equal(
+    data.frame(yn = c(1, 1)) |>
+      tbl_summary() |>
+      getElement("inputs") |>
+      getElement("type") |>
+      getElement("yn"),
+    "categorical"
+  )
+
+
+})
+
+test_that("tbl_summary(type) proper errors/messages", {
+  # grade cannot be summarized continuously, and we'll see reports in the console
+  expect_snapshot(
+    tbl <- tbl_summary(
+      trial,
+      include = grade,
+      type = grade ~ "continuous"
+    )
+  )
+  expect_equal(tbl$table_body$stat_0, "NA (NA, NA)")
+
+  # unobserved levels cannot be summarized for a dichotomous summary
+  expect_snapshot(
+    error = TRUE,
+    tbl_summary(
+      trial,
+      include = grade,
+      type = grade ~ "dichotomous",
+      value = grade ~ "IV"
+    )
+  )
+
+  # error when no clear dichotomous value present
+  expect_snapshot(
+    error = TRUE,
+    tbl_summary(
+      trial,
+      include = grade,
+      type = grade ~ "dichotomous"
+    )
+  )
+})
+
+# tbl_summary(value) -----------------------------------------------------------
 test_that("tbl_summary(value)", {
   # ensure grade is coerced to dichotomous and response defaults to dichotomous
   expect_error(
@@ -163,7 +380,65 @@ test_that("tbl_summary(value) errors properly", {
   )
 })
 
+# tbl_summary(missing) ---------------------------------------------------------
+test_that("tbl_summary(missing)", {
+  # default is correctly "ifany"
+  expect_equal(
+    tbl_summary(
+      trial,
+      include = c(trt, age)
+    ) |>
+      as.data.frame(),
+    tbl <- tbl_summary(
+      trial,
+      include = c(trt, age),
+      missing = "ifany"
+    ) |>
+      as.data.frame()
+  )
+  # age includes an Unknown row, and trt does not
+  expect_equal(tbl[,1], c("Chemotherapy Treatment", "Drug A", "Drug B", "Age", "Unknown"))
 
+  # all vars have a missing row when requested
+  expect_equal(
+    tbl_summary(
+      trial,
+      include = c(trt, age),
+      missing = "always"
+    ) |>
+      getElement("table_body") |>
+      dplyr::filter(row_type %in% "missing") |>
+      nrow(),
+    2L
+  )
+
+  # None of the vars have a missing row when requested
+  expect_equal(
+    tbl_summary(
+      trial,
+      include = c(trt, age),
+      missing = "no"
+    ) |>
+      getElement("table_body") |>
+      dplyr::filter(row_type %in% "missing") |>
+      nrow(),
+    0L
+  )
+
+  expect_snapshot(
+    error = TRUE,
+    tbl_summary(
+      trial,
+      missing = "NOT AN OPTION"
+    )
+  )
+})
+
+# tbl_summary(missing_text) ----------------------------------------------------
+
+
+
+# tbl_summary(sort) ------------------------------------------------------------
 test_that("tbl_summary(sort)", {
   expect_equal(
     tbl_summary(mtcars, sort = all_categorical() ~ "frequency", include = cyl) |>